Refactor: Git as SCM plugin
