+++ dvc/temp_git_repo.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 import os
 import shutil
 import tempfile
